Tests for ReportError.
