+++ test/NuGet.Clients.Tests/NuGet.VisualStudio.Common.Test/OutputConsoleLoggerTests.Log.cs
@@ -3,7 +3,6 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using FluentAssertions;
 using Moq;
 using NuGet.Common;
 using Xunit;
@@ -12,7 +11,7 @@ namespace NuGet.VisualStudio.Common.Test
 {
     public partial class OutputConsoleLoggerTests
     {
-        public class Log : OutputConsoleLoggerTests
+        public class Log : LogAndReportErrorTests
         {
             [Fact]
             public void Gets_MSBuild_verbosity_from_shell()
@@ -74,73 +73,14 @@ namespace NuGet.VisualStudio.Common.Test
             [MemberData(nameof(GetMessageVariationsWhichAreReported))]
             public void Adds_entry_to_error_list(LogLevel logLevel, int verbosityLevel)
             {
-                ErrorListTableEntry[] errorListTableEntries = null;
-
-                _errorList.Reset();
-                _errorList.Setup(el => el.AddNuGetEntries(It.IsAny<ErrorListTableEntry[]>()))
-                          .Callback<ErrorListTableEntry[]>(elte => errorListTableEntries = elte);
-
-                _msBuildOutputVerbosity = verbosityLevel;
-
-                _outputConsoleLogger.Log(new LogMessage(logLevel, "message"));
-
-                _errorList.Verify(el => el.AddNuGetEntries(It.IsAny<ErrorListTableEntry[]>()));
-
-                errorListTableEntries.Length.Should().Be(1);
-                errorListTableEntries[0].Message.Message.Should().Be("message");
-                errorListTableEntries[0].Message.Level.Should().Be(logLevel);
-            }
-
-            public static IEnumerable<object[]> GetMessageVariationsWhichAreReported()
-            {
-                return GetMessageVariations()
-                      .Where(v => v.reported)
-                      .Select(v => v.variation);
+                VerifyThatEntryToErrorListIsAdded(_outputConsoleLogger.Log, logLevel, verbosityLevel);
             }
 
             [Theory]
             [MemberData(nameof(GetMessageVariationsWhichAreNotReported))]
             public void Does_not_add_entry_to_error_list(LogLevel logLevel, int verbosityLevel)
             {
-                _errorList.Reset();
-                _msBuildOutputVerbosity = verbosityLevel;
-
-                _outputConsoleLogger.Log(new LogMessage(logLevel, "message"));
-
-                _errorList.VerifyNoOtherCalls();
-            }
-
-            public static IEnumerable<object[]> GetMessageVariationsWhichAreNotReported()
-            {
-                return GetMessageVariations()
-                      .Where(v => !v.reported)
-                      .Select(v => v.variation);
-            }
-
-            private static IEnumerable<(bool logged, bool reported, object[] variation)> GetMessageVariations()
-            {
-                foreach (var verbosityLevel in new int[] { 0, 1, 2, 3, 4 })
-                {
-                    // Every information message is logged, but not reported, regardless of verbosity level.
-                    yield return (logged: true, reported: false, new object[] { LogLevel.Information, verbosityLevel });
-
-                    // Every warning and error message is logged and reported, regardless of verbosity level.
-                    yield return (logged: true, reported: true, new object[] { LogLevel.Warning, verbosityLevel });
-                    yield return (logged: true, reported: true, new object[] { LogLevel.Error, verbosityLevel });
-
-                    // For all others log levels, we are logging, but not reporting, only if verbosity level is above 2
-                    foreach (var logLevel in new LogLevel[] { LogLevel.Minimal, LogLevel.Verbose, LogLevel.Debug })
-                    {
-                        if (verbosityLevel > 2)
-                        {
-                            yield return (logged: true, reported: false, new object[] { logLevel, verbosityLevel });
-                        }
-                        else
-                        {
-                            yield return (logged: false, reported: false, new object[] { logLevel, verbosityLevel });
-                        }
-                    }
-                }
+                VerifyThatEntryToErrorListIsNotAdded(_outputConsoleLogger.Log, logLevel, verbosityLevel);
             }
         }
     }
